+++ src/components/SelectionList/Search/TextWithIconCell.tsx
